Small edits to settings and Number input for validation and fix a bug where on change of the population density incorrect figures were showing in the legend

+++ app/portfolio/game-of-life/GameOfLife/components/SettingsPanel/NumberInput/NumberInput.tsx
@@ -70,7 +70,7 @@ const NumberInput: React.FC<{inputSettings: NumberInputSettings}> = ({inputSetti
     const isInvalid = GameOfLifeUtil.numberIsValid(
       inputSettings.min, 
       inputSettings.max, 
-      parseInt(inputCurrent.value)
+      parseFloat(inputCurrent.value)
     );  
 
     if(isInvalid) {
